fix(cat): keep queue chrome mounted when filters change
Move search, filter, download, and bulk actions into the CAT header so a queue refetch only skeletons the list instead of replacing the whole workspace.
