+++ apps/hyperlocalise-web/src/components/cat/side-by-side/cat-side-by-side-panel.tsx
@@ -12,30 +12,16 @@
  * of this software will be governed by the GNU General Public License
  * Version 2.0 or later.
  */
-import { FilterIcon, SearchIcon } from "@hugeicons/core-free-icons";
-import { HugeiconsIcon } from "@hugeicons/react";
-import { DownloadIcon } from "lucide-react";
 import { observer } from "mobx-react-lite";
 import { useCallback, useMemo } from "react";
-import { FormattedMessage, useIntl } from "react-intl";
+import { FormattedMessage } from "react-intl";
 
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Input } from "@/components/ui/input";
 import { Spinner } from "@/components/ui/spinner";
 import { cn } from "@/lib/primitives/cn";
 
 import { CatQueueSkeletonList } from "@/components/cat/queue/cat-queue-skeleton-list";
-import { catQueueFilterValues, type CatQueueFilter } from "@/components/cat/queue/cat-queue-filter";
+import type { CatQueueFilter } from "@/components/cat/queue/cat-queue-filter";
 import type { CatQueuePagination } from "@/components/cat/queue/cat-queue-panel";
 import {
   catQueuePanelMessages,
@@ -50,23 +36,9 @@ import type {
   CatTranslationMemoryMatch,
 } from "@/components/cat/shared/types";
 import { useCatWorkspace } from "@/components/cat/workspace/cat-workspace-context";
-import { CatWorkspaceViewSwitcherConnected } from "@/components/cat/workspace/cat-workspace-view-switcher-connected";
 
 import { CatSideBySideIntelligencePanel } from "./cat-side-by-side-intelligence-panel";
 import { CatSideBySideVirtualList } from "./cat-side-by-side-virtual-list";
-
-const queueFilterMessageByValue: Record<
-  CatQueueFilter,
-  (typeof catQueuePanelMessages)[keyof typeof catQueuePanelMessages]
-> = {
-  all: catQueuePanelMessages.filterAll,
-  untranslated: catQueuePanelMessages.filterUntranslated,
-  needs_review: catQueuePanelMessages.filterNeedsReview,
-  reviewed: catQueuePanelMessages.filterReviewed,
-  has_issues: catQueuePanelMessages.filterHasIssues,
-  skipped: catQueuePanelMessages.filterSkipped,
-  hidden: catQueuePanelMessages.filterHidden,
-};
 
 export const CatSideBySidePanel = observer(function CatSideBySidePanel({
   segments,
@@ -101,18 +73,12 @@ export const CatSideBySidePanel = observer(function CatSideBySidePanel({
   showVisualContext,
   canLookupFreshContext,
   search = "",
-  onSearchChange,
-  isSearching = false,
   queueFilter = "all",
-  onQueueFilterChange,
-  availableQueueFilters = catQueueFilterValues,
   isFetchingPage = false,
   isQueueLoading = false,
   pagination = null,
   hasMoreQueue = false,
   onLoadMoreQueue,
-  onDownloadFilteredView,
-  isDownloadingFilteredView = false,
   onFocusSegment,
   onTargetChange,
   onApprove,
@@ -172,18 +138,12 @@ export const CatSideBySidePanel = observer(function CatSideBySidePanel({
   showVisualContext: boolean;
   canLookupFreshContext: boolean;
   search?: string;
-  onSearchChange?: (value: string) => void;
-  isSearching?: boolean;
   queueFilter?: CatQueueFilter;
-  onQueueFilterChange?: (filter: CatQueueFilter) => void;
-  availableQueueFilters?: CatQueueFilter[];
   isFetchingPage?: boolean;
   isQueueLoading?: boolean;
   pagination?: CatQueuePagination | null;
   hasMoreQueue?: boolean;
   onLoadMoreQueue?: () => void;
-  onDownloadFilteredView?: (format: "csv" | "tmx" | "xlf" | "xliff") => void;
-  isDownloadingFilteredView?: boolean;
   onFocusSegment: (segmentId: string) => void;
   onTargetChange: (segmentId: string, value: string) => void;
   onApprove?: (segmentId: string) => void;
@@ -220,7 +180,6 @@ export const CatSideBySidePanel = observer(function CatSideBySidePanel({
   segmentShareUrl?: string | null;
   className?: string;
 }) {
-  const intl = useIntl();
   const store = useCatWorkspace();
   const hoveredSegmentId = store.ui.hoveredSegmentId;
   const intelligenceSegmentId = store.intelligenceSegmentId;
@@ -256,118 +215,7 @@ export const CatSideBySidePanel = observer(function CatSideBySidePanel({
       )}
     >
       <div className="flex h-full min-h-0 min-w-0 flex-col overflow-hidden">
-        <div className="shrink-0 space-y-3 border-b border-border px-4 py-3">
-          <div className="flex items-center justify-between gap-3">
-            {onSearchChange ? (
-              <div className="relative min-w-0 flex-1">
-                <HugeiconsIcon
-                  icon={SearchIcon}
-                  className="pointer-events-none absolute top-1/2 left-2.5 size-4 -translate-y-1/2 text-muted-foreground"
-                />
-                <Input
-                  value={search}
-                  onChange={(event) => onSearchChange(event.target.value)}
-                  placeholder={intl.formatMessage(catQueuePanelMessages.searchPlaceholder)}
-                  aria-label={intl.formatMessage(catQueuePanelMessages.searchAria)}
-                  className="h-9 pl-9"
-                />
-                {isSearching ? (
-                  <Spinner className="absolute top-1/2 right-2.5 size-4 -translate-y-1/2" />
-                ) : null}
-              </div>
-            ) : (
-              <div className="flex-1" />
-            )}
-
-            <div className="flex shrink-0 items-center gap-2">
-              {onQueueFilterChange ? (
-                <DropdownMenu>
-                  <DropdownMenuTrigger
-                    render={
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        className="h-9 gap-1.5 px-2.5"
-                        aria-label={intl.formatMessage(catQueuePanelMessages.filterQueueAria)}
-                      />
-                    }
-                  >
-                    <HugeiconsIcon icon={FilterIcon} className="size-4" />
-                    <span className="hidden text-xs sm:inline">
-                      <FormattedMessage {...queueFilterMessageByValue[queueFilter]} />
-                    </span>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent align="end" className="min-w-44">
-                    <DropdownMenuRadioGroup
-                      value={queueFilter}
-                      onValueChange={(value) => {
-                        if (
-                          value === "all" ||
-                          value === "untranslated" ||
-                          value === "needs_review" ||
-                          value === "reviewed" ||
-                          value === "has_issues" ||
-                          value === "skipped"
-                        ) {
-                          onQueueFilterChange(value);
-                        }
-                      }}
-                    >
-                      {availableQueueFilters.map((filter) => (
-                        <DropdownMenuRadioItem key={filter} value={filter}>
-                          <FormattedMessage {...queueFilterMessageByValue[filter]} />
-                        </DropdownMenuRadioItem>
-                      ))}
-                    </DropdownMenuRadioGroup>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              ) : null}
-
-              {onDownloadFilteredView ? (
-                <DropdownMenu>
-                  <DropdownMenuTrigger
-                    render={
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                        className="h-9 gap-1.5 px-2.5"
-                        disabled={isDownloadingFilteredView}
-                        aria-label={intl.formatMessage(catQueuePanelMessages.downloadFilteredAria)}
-                      />
-                    }
-                  >
-                    {isDownloadingFilteredView ? (
-                      <Spinner className="size-4" />
-                    ) : (
-                      <DownloadIcon className="size-4" />
-                    )}
-                    <span className="hidden text-xs sm:inline">
-                      <FormattedMessage {...catQueuePanelMessages.downloadFiltered} />
-                    </span>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent align="end" className="w-40">
-                    <DropdownMenuGroup>
-                      <DropdownMenuLabel>
-                        <FormattedMessage {...catQueuePanelMessages.downloadFilteredFormatLabel} />
-                      </DropdownMenuLabel>
-                      {(["csv", "tmx", "xlf", "xliff"] as const).map((format) => (
-                        <DropdownMenuItem
-                          key={format}
-                          onClick={() => onDownloadFilteredView(format)}
-                        >
-                          {format.toUpperCase()}
-                        </DropdownMenuItem>
-                      ))}
-                    </DropdownMenuGroup>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              ) : null}
-
-              <CatWorkspaceViewSwitcherConnected />
-            </div>
-          </div>
-
+        <div className="shrink-0 border-b border-border px-4 py-3">
           <div className="grid grid-cols-2 gap-0 text-xs font-medium tracking-wide text-muted-foreground uppercase">
             <p className="border-r border-border pr-4">
               <FormattedMessage {...catSideBySidePanelMessages.sourceColumn} />
