+++ apps/hyperlocalise-web/src/components/cat/workspace/store/cat-domain-stores.test.ts
@@ -78,6 +78,26 @@ describe("CatQueueStore", () => {
     expect(queue.checkedSegmentIds.size).toBe(0);
   });
 
+  it("stores queue search independently of the segment list", () => {
+    const queue = new CatQueueStore();
+
+    queue.setSearch("welcome");
+
+    expect(queue.search).toBe("welcome");
+  });
+
+  it("clears checked segments when bulk selection mode is turned off", () => {
+    const queue = new CatQueueStore();
+    queue.replace([...queueSegments]);
+    queue.toggleChecked("seg-02", true);
+
+    queue.setSelectionMode(true);
+    queue.setSelectionMode(false);
+
+    expect(queue.selectionMode).toBe(false);
+    expect(queue.checkedSegmentIds.size).toBe(0);
+  });
+
   it("toggles hidden metadata on queue segments", () => {
     const queue = new CatQueueStore();
     queue.replace([...queueSegments]);
